% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_colour_datalab_c}
\alias{scale_colour_datalab_c}
\alias{scale_color_datalab_c}
\title{Datalab continuous colour scale}
\usage{
scale_colour_datalab_c(..., reverse = FALSE, na.value = "grey80")

scale_color_datalab_c(..., reverse = FALSE, na.value = "grey80")
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn()}}.}

\item{reverse}{Logical. Reverse palette order.}

\item{na.value}{Colour for missing values.}
}
\description{
Datalab continuous colour scale
}
