% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-discrete.R
\name{scale_colour_datalab_d}
\alias{scale_colour_datalab_d}
\alias{scale_color_datalab_d}
\title{Datalab discrete colour scale}
\usage{
scale_colour_datalab_d(..., reverse = FALSE)

scale_color_datalab_d(..., reverse = FALSE)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{reverse}{Logical. Reverse palette order.}
}
\description{
Datalab discrete colour scale
}
