% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_fill_datalab_c}
\alias{scale_fill_datalab_c}
\title{Datalab continuous fill scale}
\usage{
scale_fill_datalab_c(..., reverse = FALSE, na.value = "grey80")
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}.}

\item{reverse}{Logical. Reverse palette order.}

\item{na.value}{Colour for missing values.}
}
\description{
Datalab continuous fill scale
}
