% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-discrete.R
\name{scale_fill_datalab_d}
\alias{scale_fill_datalab_d}
\title{Datalab discrete fill scale}
\usage{
scale_fill_datalab_d(..., reverse = FALSE)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{reverse}{Logical. Reverse palette order.}
}
\description{
Datalab discrete fill scale
}
