% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_datalab}
\alias{theme_datalab}
\title{Datalab ggplot2 theme}
\usage{
theme_datalab(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
Datalab ggplot2 theme
}
